requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg[binary]
dj-database-url
python-decouple
celery
redis
pillow
requests
openai
anthropic
python-dotenv
gunicorn
cryptography
qrcode
whitenoise
google-generativeai
pymupdf
google-genai
